Senior project
The project is intended to develop a mobile application which is capable of scanning QR code
and generate one. This application will be used by the people that want to apply for job
(meeting in person). Users can log into the app and scan QR code for different user.  Every
QR code the user scan is stored in an array and can be viewed in My wallet view controller.
We can edit/delete any of stored data in this view. - Home View : provided 3 features:
   Qr code scan, Account Info, My Wallet. The next view is QR code View : A view which opens
camera and scan for QR code.